Initial Commit - basic back-up on a 5 minute interval
I've been wanting to tackle this project for a long time, seeing how often Excel crashes or corrupts my files and I have nothing to fall back on.

I'm sure this will be a difficult untaking to "do well", but I have to start somewhere. This bare-bones add-in should start auto-saving any workbook opened in the same instance to a sister directory (of the opened workbook), creating a new copy every 5 minutes.

Much room for improvement, but at least this basic functionality appears to be working.

Currently have it set up to save once initially, then again every 5 minutes if (and only if) a change to the workbook has been detected since the last save. That should avoid filling up my hard drive if I leave Excel open overnight.

That works fine unless the workbook itself contains a timer (e.g. which updates a data connection every 5 minutes), which results in a change being detected. As such, I further limit the auto-recovery folder to the 10 most recent saves.

In the end, all this behaviour should be configurable via a little pop-up dialog or similar.